feat: redux 대신 zustand 사용
