+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^2.2.7",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -14,13 +13,12 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-modal-sheet": "^3.1.0",
-    "react-redux": "^9.1.2",
     "react-router-dom": "^6.26.1",
     "react-scripts": "5.0.1",
-    "redux": "^5.0.1",
     "styled-components": "^6.1.12",
     "typescript": "^4.9.5",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "zustand": "^4.5.5"
   },
   "scripts": {
     "start": "react-scripts start",
