+++ src/index.tsx
@@ -5,18 +5,14 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 // import { ThemeProvider } from "styled-components";
 import GlobalStyle from "./globalStyles";
-import { Provider } from "react-redux";
-import store from "./store";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 root.render(
   <React.StrictMode>
-    <Provider store={store}>
-      <GlobalStyle />
-      <App />
-    </Provider>
+    <GlobalStyle />
+    <App />
   </React.StrictMode>
 );
