--- src/store/index.ts
@@ -1,13 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-import postsReducer from "./postsSlice";
-
-const store = configureStore({
-  reducer: {
-    posts: postsReducer,
-  },
-});
-
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
-
-export default store;
